Use a local label to refer to `stack_init_trampoline_return` in asm
This avoids issues with relocations when building as a shared library.
Specifically, `stack_init_trampoline_return` being a global symbol means
that it could be overriden and must therefore always be accessed via the
GOT. Using a local label avoids this issue and allows the use of
relative addressing.

Fixes wasmerio/wasmer#5721